// src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
      chapters: {
        Row: {
          course_id: string
          created_at: string
          description: string | null
          id: string
          serial_number: number
          sub_title: string | null
          title: string
          updated_at: string
        }
        Insert: {
          course_id: string
          created_at?: string
          description?: string | null
          id?: string
          serial_number?: number
          sub_title?: string | null
          title: string
          updated_at?: string
        }
        Update: {
          course_id?: string
          created_at?: string
          description?: string | null
          id?: string
          serial_number?: number
          sub_title?: string | null
          title?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "chapters_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      courses: {
        Row: {
          course_language: number | null
          course_language_dialect: number | null
          created_at: string
          description: string | null
          description_html: string | null
          description_json: string | null
          id: string
          is_published: boolean | null
          language_taught: number | null
          language_taught_dialect: number | null
          profile_id: string
          published_at: string | null
          short_description: string | null
          support_link: string | null
          thumbnail: string | null
          title: string
          updated_at: string
        }
        Insert: {
          course_language?: number | null
          course_language_dialect?: number | null
          created_at?: string
          description?: string | null
          description_html?: string | null
          description_json?: string | null
          id?: string
          is_published?: boolean | null
          language_taught?: number | null
          language_taught_dialect?: number | null
          profile_id: string
          published_at?: string | null
          short_description?: string | null
          support_link?: string | null
          thumbnail?: string | null
          title: string
          updated_at?: string
        }
        Update: {
          course_language?: number | null
          course_language_dialect?: number | null
          created_at?: string
          description?: string | null
          description_html?: string | null
          description_json?: string | null
          id?: string
          is_published?: boolean | null
          language_taught?: number | null
          language_taught_dialect?: number | null
          profile_id?: string
          published_at?: string | null
          short_description?: string | null
          support_link?: string | null
          thumbnail?: string | null
          title?: string
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "courses_course_language_dialect_fkey"
            columns: ["course_language_dialect"]
            isOneToOne: false
            referencedRelation: "dialects"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "courses_course_language_fkey"
            columns: ["course_language"]
            isOneToOne: false
            referencedRelation: "languages"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "courses_language_taught_dialect_fkey"
            columns: ["language_taught_dialect"]
            isOneToOne: false
            referencedRelation: "dialects"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "courses_language_taught_fkey"
            columns: ["language_taught"]
            isOneToOne: false
            referencedRelation: "languages"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "courses_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      languages: {
        Row: {
          code: string
          created_at: string
          id: number
          name_en: string
          name_jpn: string | null
          name_native: string | null
          updated_at: string
        }
        Insert: {
          code: string
          created_at?: string
          id?: number
          name_en: string
          name_jpn?: string | null
          name_native?: string | null
          updated_at?: string
        }
        Update: {
          code?: string
          created_at?: string
          id?: number
          name_en?: string
          name_jpn?: string | null
          name_native?: string | null
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}